+++ site/homepage/content/release-notes.md
@@ -4,7 +4,7 @@ title = "What is new & noteworthy in Hono?"
 description = "Information about changes in recent Hono releases. Includes new features, fixes, enhancements and API changes."
 +++
 
-## 1.13.0 (not yet released)
+## 2.0.0 (not yet released)
 
 ### Fixes & Enhancements
 
@@ -13,6 +13,11 @@ description = "Information about changes in recent Hono releases. Includes new f
   *hono.mongodb.connectionString* property. This has been fixed.
 * When a tenant or device gets disabled or deleted, any open AMQP or MQTT connections from clients having authenticated
   themselves as belonging to that tenant or device are getting closed now.
+
+# API Changes
+
+* The MQTT adapter no longer accepts the `+` wild card character for the device ID in a topic filter used for
+  subscribing to commands.
 
 ## 1.12.0